+++ programs/benches/map.c
@@ -4,7 +4,7 @@
 
 typedef struct map_return_values
 {
-    uint64_t remaining_gas;
+    unsigned __int128 remaining_gas;
     struct {
         uint8_t discriminant;
         union {
